edit course is done (sort of)
if you want to access it, go to user's profile. there you'll see courses with edit option (atm it's showing all the courses in database which we will fix later). the problem is it won't open up the page that you want to edit to. e.g a user wants to edit RAD course but instead, it opens up APT course. we need to figure out how to put current course in view link.